wip: api wb scheduler
